keep remsql communication protocol encapsulated; remove remcur
